mainlist: sanity check going back from top menu
When going back to the top directory, the code passes "...". But what
if you go back *from* the top directory?

In this case, the menu goes crazy and loads the SD card directory, and
it is impossible to go back to the top directory.

At the top directory, the location is not "..." but actually "". It is
also possible for the directory passed to be ".../". Add a check for
all these probabilities and treat them like the top directory.